+++ graph/network/distance.go
@@ -18,9 +18,9 @@ import (
 //
 // For directed graphs the incoming paths are used. Infinite distances are
 // not considered.
-func Closeness(g graph.Graph, p path.AllShortest) map[int]float64 {
+func Closeness(g graph.Graph, p path.AllShortest) map[int64]float64 {
 	nodes := g.Nodes()
-	c := make(map[int]float64, len(nodes))
+	c := make(map[int64]float64, len(nodes))
 	for _, u := range nodes {
 		var sum float64
 		for _, v := range nodes {
@@ -45,9 +45,9 @@ func Closeness(g graph.Graph, p path.AllShortest) map[int]float64 {
 //
 // For directed graphs the incoming paths are used. Infinite distances are
 // not considered.
-func Farness(g graph.Graph, p path.AllShortest) map[int]float64 {
+func Farness(g graph.Graph, p path.AllShortest) map[int64]float64 {
 	nodes := g.Nodes()
-	f := make(map[int]float64, len(nodes))
+	f := make(map[int64]float64, len(nodes))
 	for _, u := range nodes {
 		var sum float64
 		for _, v := range nodes {
@@ -72,9 +72,9 @@ func Farness(g graph.Graph, p path.AllShortest) map[int]float64 {
 //
 // For directed graphs the incoming paths are used. Infinite distances are
 // not considered.
-func Harmonic(g graph.Graph, p path.AllShortest) map[int]float64 {
+func Harmonic(g graph.Graph, p path.AllShortest) map[int64]float64 {
 	nodes := g.Nodes()
-	h := make(map[int]float64, len(nodes))
+	h := make(map[int64]float64, len(nodes))
 	for i, u := range nodes {
 		var sum float64
 		for j, v := range nodes {
@@ -101,9 +101,9 @@ func Harmonic(g graph.Graph, p path.AllShortest) map[int]float64 {
 //
 // For directed graphs the incoming paths are used. Infinite distances are
 // not considered.
-func Residual(g graph.Graph, p path.AllShortest) map[int]float64 {
+func Residual(g graph.Graph, p path.AllShortest) map[int64]float64 {
 	nodes := g.Nodes()
-	r := make(map[int]float64, len(nodes))
+	r := make(map[int64]float64, len(nodes))
 	for i, u := range nodes {
 		var sum float64
 		for j, v := range nodes {
